refactor: move communicator to distributed/ subpackage to align with vllm-plugin-FL structure
- sglang_fl/communicator.py → sglang_fl/distributed/communicator.py
- sglang_fl/flagcx_communicator.py → sglang_fl/distributed/device_communicators/flagcx.py
- Update all import references (3 places)
- Add distributed/__init__.py with re-export
